Crowbar 0.49 - 2017-10-14_0.49
Changelog:

=== General ===

    If a folder is drag-dropped onto a tab that can accept it, then set up that tab and stay on that tab.
    Updated the font to match what Windows uses instead of using the default that Visual Studio assigns.
    "Browse" buttons will now open the right-most folder in the path that exists if the full path does not exist.

=== Unpack ===

    New Unpack tab. [Completely revised from any previous versions that had it.]
    Unpack all or some files from one or more VPK files via buttons or drag-n-drop.
    Unpacks as quickly as GCFScape.
    Find feature that stores the results such that the files in the results can be unpacked and searched.

=== Decompile ===

    FIX: For Vampire: The Masquerade - Bloodlines models, UV map is upside-down.
    FIX: Decompiling MDL v10 uses same syntax as Source models, but it should not.
    Decompiling an MDL v10 should be the same or better than "Half-Life MDL Decompiler 1.2" (mdldec) by Kratisto.
    During testing, Crowbar's values for root-bone movement (walkframe info) matched a lot more closely to original values.
    Also, Crowbar decompiles $texrendermode data.
    FIX: Decompiling an MDL v10 without its sequence group MDL files or without its texture MDL file stops the decompile without proper messages.
    FIX: After decompiling an MDL v10, the bottom MDL combobox is not filled-in with the MDL file name, even on successful compile.
    FIX: Decompile tab "Use All in Compile" button does not set the "QC input" to full path in Compile tab.
    Changed file input and output areas.
    Split the Options group box into two boxes to help clarify the meaning of the options.
    SMD files are now named to use the original names of meshes while avoiding SMD file name conflicts.
    This includes fixing the reported bug where the Opposing Force "massn.mdl" decompiled only one "frontkick" animation SMD file.
    Added new option: "Only changed materials in $texturegroup lines".
    Write batch decompile log to just one file.

=== Compile ===

    FIX: Using the Subfolder Output option, when a QC file has a comment on the same line after $modelname "file.mdl", fails to place compiled files in Subfolder.
    FIX: Compiling MDL v10 fails.
    Changed file input and output areas.
    Compile Options now change based on whether game is GoldSource or Source.
    Preserve the folder tree that studiomdl makes for the model files for all Output Folder options. Removed the "Folder for each model" option because it is no longer needed.
    Provides more details about Crowbar's managing of folders and files in the Compile log.
    Write batch compile log to just one file.

=== Preview and View ===

    FIX: The "crowbar" folder is created and "liblist.gam" file modified for GoldSource engine game, but should not be done.
    FIX: The "materials" folder is copied for "View" and "View as Replacement", but should not be if model files are in a game's "models" folder tree.
    FIX: Auto open of MDL file (when its related files are missing) does not update Info.

=== Options ===

    Added Unpack options.
    Added an Apply box that appears if Windows is not using what is specified.
    Added options for how to set up Crowbar when a folder is dragged onto "Crowbar.exe" file.

=== Help ===

    New Help tab.

=== About ===

    Added steam group link under Crowbar to make it more obvious.
    Added steam profile link under ZeqMacaw to make it more obvious.
    Added "Unpack VPK files" to the list of tools in the toolset.

=== Set Up Games ===

    Sven Co-op: Set the correct paths for model compiler and mapping tool from the Sven Co-op SDK.
    Added "Alien Swarm: Reactive Drop" game setup to default list.

+++ Crowbar/Core/SourceModel/SourceModel32/SourceMdlFileData/SourceMdlFileData32.vb
@@ -245,7 +245,7 @@
 	Public thePoseParamDescs As List(Of SourceMdlPoseParamDesc)
 	Public theSequenceDescs As List(Of SourceMdlSequenceDesc32)
 	Public theSequenceGroups As List(Of SourceMdlSequenceGroup37)
-	Public theSkinFamilies As List(Of List(Of Integer))
+	Public theSkinFamilies As List(Of List(Of Short))
 	Public theSurfacePropName As String
 	Public theTexturePaths As List(Of String)
 	Public theTextures As List(Of SourceMdlTexture37)
